Put VotingOptionCard component into its own page
