+++ client/src/pages/Room/RoomPage.jsx
@@ -19,6 +19,7 @@ import AddIcon from '@mui/icons-material/Add';
 import RemoveIcon from '@mui/icons-material/Remove';
 import ArrowBackOutlinedIcon from '@mui/icons-material/ArrowBackOutlined';
 import PermanentDrawerLeft from './Drawer';
+import VotingOptionCard from './VotingOptionCard';
 
 const dummyVotingOptions = [
   {
@@ -84,47 +85,6 @@ const Room = ({}) => {
   //Navigates back to the login page.
   const navigateBack = () => {
     navigate('/');
-  };
-
-  const VotingOptionCard = ({
-    name,
-    votesIn,
-    userVotesIn,
-    totalAvailableVotes,
-  }) => {
-    return (
-      <Card
-        style={{
-          padding: '1rem',
-          display: 'flex',
-          gap: '1rem',
-          alignItems: 'center',
-          margin: '1rem',
-        }}
-      >
-        <Typography
-          textAlign={'left'}
-          width={'45%'}
-          textOverflow={'ellipsis'}
-          overflow={'hidden'}
-        >
-          {name}
-        </Typography>
-        <Chip label={`${votesIn} / ${totalAvailableVotes}`} />
-        <Chip
-          label={`${userVotesIn}`}
-          sx={{ background: '#B7CFEE', color: '#2E419D' }}
-        />
-        <Box>
-          <IconButton>
-            <AddIcon sx={{ color: 'green' }} />
-          </IconButton>
-          <IconButton>
-            <RemoveIcon sx={{ color: 'red' }} />
-          </IconButton>
-        </Box>
-      </Card>
-    );
   };
 
   return (
